+++ frontend/src/pages/Login.jsx
@@ -7,7 +7,6 @@ const Login = ({}) => {
 
   return (
     <div className="flex min-h-full items-center justify-center py-12 px-4 sm:px-6 lg:px-8 font-sans">
-      <button onClick={() => toast.succes('hello')}>hello</button>
       <div className="w-full max-w-md space-y-8">
         <div>
           <h2 className="mt-6 text-center text-2xl uppercase tracking-tight text-gray-900">
